# src/main/java/ru/devhub/web/web/reference/type/TypeDto.java
package ru.devhub.web.web.reference.type;

import java.util.UUID;

public class TypeDto {
    private UUID id;
    private String name;

    public UUID getId() {
        return id;
    }

    public void setId(UUID id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}