# src/main/resources/openapi.yaml
openapi: "3.1.0"

info:
  title: DevHub API
  description: |
    MVP платформы проектно-центричного обучения IT-специалистов.
    Все эндпоинты требуют Bearer JWT (Keycloak). Claim `business_id` содержит UUID пользователя.
  version: "0.1.0"
  contact:
    name: DevHub Team

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

security:
  - bearerAuth: []

tags:
  - name: projects
    description: Управление проектами
  - name: reference
    description: Справочники (роли, технологии, типы проектов)

# ─────────────────────────────────────────────────────────────
paths:

  # ── PROJECTS ─────────────────────────────────────────────
  /projects:
    get:
      tags: [ projects ]
      summary: Список проектов (paginated)
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        "200":
          description: Страница проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [ projects ]
      summary: Создать проект
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          description: Проект создан
          headers:
            Location:
              description: URI созданного ресурса
              schema:
                type: string
                example: /api/v1/projects/550e8400-e29b-41d4-a716-446655440000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'

    get:
      tags: [ projects ]
      summary: Получить проект по ID
      operationId: getProject
      responses:
        "200":
          description: Детали проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [ projects ]
      summary: Обновить проект (только владелец)
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "200":
          description: Проект обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

  # ── REFERENCE ────────────────────────────────────────────
  /reference/roles:
    get:
      tags: [ reference ]
      summary: Список ролей участников
      operationId: listRoles
      responses:
        "200":
          description: Все доступные роли
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRolesResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'

  /reference/technologies:
    get:
      tags: [ reference ]
      summary: Список технологий
      operationId: listTechnologies
      responses:
        "200":
          description: Все доступные технологии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTechnologiesResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'

  /reference/types:
    get:
      tags: [ reference ]
      summary: Список типов проектов
      operationId: listProjectTypes
      responses:
        "200":
          description: Все доступные типы проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTypesResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'

# ─────────────────────────────────────────────────────────────
components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT. Claim `business_id` содержит UUID пользователя.

  # ── PARAMETERS ───────────────────────────────────────────
  parameters:
    ProjectIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID проекта

    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Номер страницы (0-based)

    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Размер страницы

  # ── RESPONSES ────────────────────────────────────────────
  responses:
    BadRequest:
      description: Невалидные входные данные
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            error: Bad Request
            message: "name: must not be blank"
            path: /api/v1/projects
            timestamp: "2024-01-15T10:30:00Z"

    Unauthorized:
      description: Отсутствует или невалидный JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Операция запрещена — только владелец может изменять проект
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 403
            error: Forbidden
            message: "Access denied: user is not the project owner"
            path: /api/v1/projects/550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00Z"

    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            error: Not Found
            message: "Project with id '550e8400-e29b-41d4-a716-446655440000' not found"
            path: /api/v1/projects/550e8400-e29b-41d4-a716-446655440000
            timestamp: "2024-01-15T10:30:00Z"

    UnprocessableEntity:
      description: Нарушение бизнес-правил (невалидный статус, несуществующая технология и т.п.)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 422
            error: Unprocessable Entity
            message: "Invalid project status: 'UNKNOWN'"
            path: /api/v1/projects
            timestamp: "2024-01-15T10:30:00Z"

  # ── SCHEMAS ──────────────────────────────────────────────
  schemas:

    # ── Enums ──────────────────────────────────────────────
    ProjectStatus:
      type: string
      enum: [ DRAFT, ACTIVE, RECRUITING, ARCHIVED ]
      description: |
        DRAFT      — черновик, не виден другим пользователям;
        ACTIVE     — активный, идёт разработка;
        RECRUITING — набор участников;
        ARCHIVED   — архивирован.

    MemberStatus:
      type: string
      enum: [ OWNER, ACTIVE, INVITED, LEFT, REMOVED ]
      description: Статус участника в проекте.

    # ── Shared sub-objects ─────────────────────────────────
    UserDto:
      type: object
      required: [ id, email, name ]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: Иван Иванов

    RoleDto:
      type: object
      required: [ id, name ]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Backend Developer

    TechnologyDto:
      type: object
      required: [ id, name ]
      properties:
        id:
          type: integer
          example: 3
        name:
          type: string
          example: Java

    TypeDto:
      type: object
      required: [ id, name ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Коммерческий проект

    # ── Member DTOs ────────────────────────────────────────
    MemberRequestDto:
      type: object
      description: Участник проекта в теле запроса (Create / Update).
      required: [ userId, status ]
      properties:
        userId:
          type: string
          format: uuid
          description: UUID пользователя, которого добавляем/обновляем.
        roleIds:
          type: array
          description: Список ID ролей участника.
          items:
            type: integer
          default: []
        status:
          $ref: '#/components/schemas/MemberStatus'

    MemberResponseDto:
      type: object
      description: Участник проекта в теле ответа.
      required: [ id, user, status, joinedAt ]
      properties:
        id:
          type: string
          format: uuid
          description: UUID записи ProjectMember.
        projectId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserDto'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
        status:
          $ref: '#/components/schemas/MemberStatus'
        joinedAt:
          type: string
          format: date-time
        leftAt:
          type: string
          format: date-time
          nullable: true

    # ── Request bodies ─────────────────────────────────────
    CreateProjectRequest:
      type: object
      required: [ name, status ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          example: DevHub Backend
        description:
          type: string
          maxLength: 4000
          nullable: true
          example: Полное описание проекта.
        shortDescription:
          type: string
          maxLength: 300
          nullable: true
          example: Краткое описание.
        typeId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        technologyIds:
          type: array
          items:
            type: integer
          default: []
        roleIds:
          type: array
          items:
            type: integer
          default: []
        members:
          type: array
          description: Начальный состав команды (владелец добавляется автоматически).
          items:
            $ref: '#/components/schemas/MemberRequestDto'
          default: []

    UpdateProjectRequest:
      type: object
      required: [ name, status ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 4000
          nullable: true
        shortDescription:
          type: string
          maxLength: 300
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        typeId:
          type: string
          format: uuid
          nullable: true
        technologyIds:
          type: array
          items:
            type: integer
          default: []
        roleIds:
          type: array
          items:
            type: integer
          default: []
        members:
          type: array
          description: Полная замена состава команды (full-replace semantics). Владелец не может быть удалён.
          items:
            $ref: '#/components/schemas/MemberRequestDto'
          default: []

    # ── Response bodies ────────────────────────────────────
    ProjectDetailResponse:
      type: object
      description: |
        Полная карточка проекта. Используется как ответ на POST (201), PUT (200) и GET /{id} (200).
      required: [ id, name, status, owner, createdAt ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        shortDescription:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        type:
          $ref: '#/components/schemas/TypeDto'
          nullable: true
        owner:
          $ref: '#/components/schemas/UserDto'
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyDto'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberResponseDto'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProjectListItemDto:
      type: object
      description: Карточка проекта для ленты/списка (облегчённый вариант без деталей членства).
      required: [ id, name, status, owner, createdAt ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        shortDescription:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProjectStatus'
        type:
          $ref: '#/components/schemas/TypeDto'
          nullable: true
        owner:
          $ref: '#/components/schemas/UserDto'
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyDto'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
        members:
          type: array
          description: Список участников — только UserDto (без деталей членства).
          items:
            $ref: '#/components/schemas/UserDto'
        createdAt:
          type: string
          format: date-time

    ListProjectResponse:
      type: object
      required: [ projects, total, page, size ]
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectListItemDto'
        total:
          type: integer
          format: int64
          description: Общее число проектов в системе.
        page:
          type: integer
          description: Текущая страница (0-based).
        size:
          type: integer
          description: Размер страницы.

    # ── Reference responses ────────────────────────────────
    ListRolesResponse:
      type: object
      required: [ roles ]
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'

    ListTechnologiesResponse:
      type: object
      required: [ technologies ]
      properties:
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyDto'

    ListTypesResponse:
      type: object
      required: [ types ]
      properties:
        types:
          type: array
          items:
            $ref: '#/components/schemas/TypeDto'

    # ── Error ──────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [ status, error, message, path, timestamp ]
      properties:
        status:
          type: integer
          description: HTTP-статус код.
          example: 404
        error:
          type: string
          description: Краткое название ошибки.
          example: Not Found
        message:
          type: string
          description: Детальное сообщение об ошибке.
          example: "Project with id '550e...' not found"
        path:
          type: string
          description: URI запроса, вызвавшего ошибку.
          example: /api/v1/projects/550e8400-e29b-41d4-a716-446655440000
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки (UTC).